requirements/colab-gpu.txt
transformers>=4.40
accelerate>=0.30
datasets>=2.19
peft>=0.10
bitsandbytes>=0.43
safetensors>=0.4
sentencepiece>=0.2
pyyaml>=6.0
